modify the generator for one of the cauth users tables

+++ generators/templates/add_users_table_to_cauth_db.rb
@@ -16,6 +16,8 @@ class AddUsersTableToCauthDb < CentralAuthentication::Migration
   end
 
   def self.down
-    drop_table :users
+    if table_exists?(:users)
+      drop_table :users
+    end
   end
 end
